Categories
добавлен пустой контроллер с категориями паролей

+++ mobilelab/mobilelab/Categories/NSString/NSString+MD5.m
@@ -14,7 +14,7 @@
 -(NSString *)MD5String {
     const char *cStr = [self UTF8String];
     unsigned char digest[CC_MD5_DIGEST_LENGTH];
-    CC_MD5( cStr, strlen(cStr), digest );
+    CC_MD5( cStr, (unsigned int)strlen(cStr), digest );
     
     NSMutableString *output = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
     for(int i = 0; i < CC_MD5_DIGEST_LENGTH; i++) {
